BPM-89: 질문하기 게시판에 글 등록시에 file없어도 받을 수 있게 로직 변경

+++ src/main/java/d83t/bpmbackend/domain/aggregate/community/service/QuestionBoardServiceImpl.java
@@ -64,7 +64,7 @@ public class QuestionBoardServiceImpl implements QuestionBoardService {
     @Override
     public QuestionBoardResponse createQuestionBoardArticle(User user, List<MultipartFile> files, QuestionBoardRequest questionBoardRequest) {
         //file은 최대 5개만 들어올 수 있다.
-        if (files.size() > 5) {
+        if (files != null && files.size() > 5) {
             throw new CustomException(Error.FILE_SIZE_MAX);
         }
         User findUser = userRepository.findByKakaoId(user.getKakaoId()).orElseThrow(() -> {
@@ -79,25 +79,26 @@ public class QuestionBoardServiceImpl implements QuestionBoardService {
                 .build();
 
         List<String> filePaths = new ArrayList<>();
-
-        for (MultipartFile file : files) {
-            String newName = FileUtils.createNewFileName(file.getOriginalFilename());
-            String filePath = fileDir + newName;
-            questionBoard.addQuestionBoardImage(QuestionBoardImage.builder()
-                    .originFileName(newName)
-                    .storagePathName(filePath)
-                    .questionBoard(questionBoard)
-                    .build());
-            filePaths.add(filePath);
-            if (env.equals("prod")) {
-                uploaderService.putS3(file, questionBoardPath, newName);
-            } else if (env.equals("local")) {
-                try {
-                    File localFile = new File(filePath);
-                    file.transferTo(localFile);
-                    FileUtils.removeNewFile(localFile);
-                } catch (IOException e) {
-                    e.printStackTrace();
+        if(files != null && files.size() != 0) {
+            for (MultipartFile file : files) {
+                String newName = FileUtils.createNewFileName(file.getOriginalFilename());
+                String filePath = fileDir + newName;
+                questionBoard.addQuestionBoardImage(QuestionBoardImage.builder()
+                        .originFileName(newName)
+                        .storagePathName(filePath)
+                        .questionBoard(questionBoard)
+                        .build());
+                filePaths.add(filePath);
+                if (env.equals("prod")) {
+                    uploaderService.putS3(file, questionBoardPath, newName);
+                } else if (env.equals("local")) {
+                    try {
+                        File localFile = new File(filePath);
+                        file.transferTo(localFile);
+                        FileUtils.removeNewFile(localFile);
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
                 }
             }
         }
